Re-organize abstraction constructor tests
