+++ analyzer/src/main/kotlin/com/github/cubuspl42/sigmaLang/analyzer/main.kt
@@ -13,7 +13,7 @@ fun main() {
         mainModuleName = "problem",
     )
 
-    when (val outcome = project.entryPoint.evaluateResult()) {
+    when (val outcome = project.entryPoint.valueThunk.evaluateInitial()) {
         is EvaluationError -> println("Error: $outcome")
         is EvaluationResult -> println("Result: ${outcome.value.dump()}")
     }
